[docs] clarify that sish.http(s)port flags are for overrides
the current documentation for the httpport and httpsport
flags are ambiguous and can mean that they can be used
without using http / https when the host can remain as
'localhost'.
This changeset clarifies the documentation to maintain
that the flags are only overrides to the default values
provided in the http / https flags.

Signed-off-by: shine <4771718+shinenelson@users.noreply.github.com>