Make welcome message configurable (#343)
The message displayed to the user upon connection (which used to be a
hard-coded "Press Ctrl-C to close the session.") can now be configured
using the "welcome-message" setting in the config file or the
"--welcome-message" command line argument. The default value is the
beforementioned message, so without an explicit configuration the
behavior will be the same as before.